requirements.txt
requests
PyCryptodome
bs4
lxml
